Safe mode

+++ main.py
@@ -2,10 +2,13 @@ import database
 import discord
 from discord.ext import commands, tasks
 from discord.ext.commands import has_permissions, MissingPermissions
-from datetime import datetime, timedelta, timezone
 
+from helpers import parseMaxAge, getReadableMaxAge
+
+from datetime import datetime, timedelta, timezone
 from dotenv import load_dotenv
 import os
+
 
 # Either load the env vars from a local env file (when testing stuff locally) or it's already been loaded (running on server)
 if os.path.isfile('.env'):
@@ -14,6 +17,11 @@ if os.path.isfile('.env'):
 else:
     print('Loading environment variables from server variables')
 
+
+# Safe mode allows you to run the bot and test stuff, without having actual results on the database or existing messages
+SAFEMODE = os.environ.get('SAFE_MODE', False).lower() == 'true'
+if SAFEMODE:
+    print('Safe mode is on')
 
 
 # Boilerplate
@@ -26,6 +34,7 @@ client = commands.Bot(command_prefix='~~', intents=intents)
 @client.event
 async def on_ready():
     delete.start()
+
 
 @client.command()
 @has_permissions(manage_messages=True)
@@ -44,47 +53,16 @@ async def deleteHere(ctx):
         await channel.send('can\'t recognise that time format, please enter it in days, hours, or minutes with for example `5d`, `12h`, or `10m`')
         return
 
+    if SAFEMODE:
+        await channel.send(f'SAFEMODE: would have added this channel for future auto purging with max message age of {getReadableMaxAge(maxAge)}')
+        return
+
     # Attempt to add channel to database
     result = database.addChannel(channel.id, maxAge)
     if result == 1:
         await channel.send('sorry, something went wrong :/ please try again (it will probably work this time :3 )')
     else:
-        await channel.send('oki! (the first purge at this max age will be in the next first purge cycle :3 )')
-
-
-def parseMaxAge(message: str) -> int:
-    """Returns the max age that was sent alongside the message in minutes
-    
-    If the max age could not be parsed, it returns -1
-    """
-    hoursPerDay = 24
-    minutesPerHour = 60
-    default = 7*hoursPerDay*minutesPerHour
-    split = message.split(' ')
-
-    if len(split) == 1:
-        return default
-    elif len(split) == 2:
-        time = split[1]     # Would be like '5d'
-        try:
-            amount = int(time[0:-1])        # Would be like 5
-            identifier = time[-1].lower()   # Would be like 'd'
-            match identifier:
-                case 'd':
-                    return amount*hoursPerDay*minutesPerHour
-                
-                case 'h':
-                    return amount*minutesPerHour
-
-                case 'm':
-                    return amount
-                
-                case _:
-                    return -1
-        except Exception:
-            return -1
-    else:
-        return -1
+        await channel.send(f'oki! Auto purging set up with max message age of {getReadableMaxAge(maxAge)} (the first purge in this channel will be in the next first purge cycle :3 )')
 
 
 @deleteHere.error
@@ -101,6 +79,9 @@ async def delete():
     for (channelID, maxAge) in database.getChannels():
         cutoff = now - timedelta(minutes=maxAge)
         channel = client.get_channel(channelID)
+        if SAFEMODE:
+            print(f'SAFEMODE: Would purge all messages in channel {channelID} before {cutoff} (UTC)')
+            continue
         
         print(f'Purging all messages in channel {channelID} before {cutoff} (UTC)')
         await channel.purge(limit=None, before=cutoff, reason='included in periodic channel purge', oldest_first=True)
@@ -111,6 +92,10 @@ async def delete():
 async def stopDeletion(ctx):
     """Command to deactivate auto purging in a channel"""
     channel = ctx.channel
+    if SAFEMODE:
+        await channel.send('SAFEMODE: would have removed this channel from auto purges')
+        return
+
     result = database.deleteChannel(channel.id)
     if result == 1:
         await channel.send('sorry, something went wrong :/ please try again (it will probably work this time :3 )')
@@ -137,12 +122,7 @@ async def info(ctx):
     # Find the channel if possible
     for (channelID, maxAge) in channels:
         if channelID == channel.id:
-            if (maxAge/24)/60 == (maxAge//24)//60:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge//24//60} days old')
-            elif maxAge/60 == maxAge//60:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge//60} hours old')
-            else:
-                await channel.send(f'This channel auto deletes messages that are more than {maxAge} minutes old')
+            await channel.send(f'This channel auto deletes messages that are more than {getReadableMaxAge(maxAge)} old')
             return
     
     await channel.send(f'This channel does not have auto deletion set up')
